Abgabe/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
pandas
python-jose[cryptography]
pyotp







